backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.6
django-ipware==7.0.1
django-rest-authemail==2.1.7
djangorestframework==3.15.2
idna==3.7
python-ipware==3.0.0
requests==2.32.3
sqlparse==0.5.0
urllib3==2.2.2
